Mainly documentation

+++ src/Narvalo.Core/Int64Encoder.cs
@@ -56,7 +56,7 @@ namespace Narvalo
 
         public static string ToBase25String(long value)
         {
-            Contract.Requires(value >= (long)0);
+            Contract.Requires(value >= 0L);
             Contract.Ensures(Contract.Result<string>() != null);
 
             return Encode(value, Base25Alphabet_, Base25AlphabetLength_);
@@ -64,7 +64,7 @@ namespace Narvalo
 
         public static string ToBase34String(long value)
         {
-            Contract.Requires(value >= (long)0);
+            Contract.Requires(value >= 0L);
             Contract.Ensures(Contract.Result<string>() != null);
 
             return Encode(value, Base34Alphabet_, Base34AlphabetLength_);
@@ -72,7 +72,7 @@ namespace Narvalo
 
         public static string ToBase58String(long value)
         {
-            Contract.Requires(value >= (long)0);
+            Contract.Requires(value >= 0L);
             Contract.Ensures(Contract.Result<string>() != null);
 
             return Encode(value, Base58Alphabet_, Base58AlphabetLength_);
@@ -80,7 +80,7 @@ namespace Narvalo
 
         public static string ToFlickrBase58String(long value)
         {
-            Require.GreaterThanOrEqualTo(value, (long)0, "value");
+            Require.GreaterThanOrEqualTo(value, 0L, "value");
             Contract.Ensures(Contract.Result<string>() != null);
 
             string result = String.Empty;
@@ -189,7 +189,7 @@ namespace Narvalo
 
         internal static string Encode(long value, char[] alphabet, int alphabetLength)
         {
-            Require.GreaterThanOrEqualTo(value, (long)0, "value");
+            Require.GreaterThanOrEqualTo(value, 0L, "value");
             Contract.Requires(alphabet != null);
             Contract.Requires(alphabetLength > 0);
             Contract.Ensures(Contract.Result<string>() != null);
